requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
opencv-python>=4.5.0
fastapi>=0.68.0
python-multipart>=0.0.5
uvicorn>=0.15.0
streamlit>=1.0.0
Pillow>=8.0.0
python-dotenv>=0.19.0
plotly>=5.0.0
optuna>=3.0.0
albumentations>=1.3.0
tqdm>=4.65.0
requests>=2.31.0
gdown>=4.7.1
ultralytics>=8.0.0
pillow>=9.5.0
scikit-learn>=1.0.0
pydantic>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
